modifications vers une installation generaliste
i think to replace a lot of things into an system for use the project
because, that have need a lot of things and I'm sure it's better if I
start to take in charge: arch and debian ...

but I want do more: an esasy to deploy web meteo station with an nginx
auto configured.

and other, I go base this on automatic scripts.
first.